requirements.txt
onnxruntime
numpy
Pillow
flask
waitress